+++ src/main/java/com/drmangotea/tfmg/content/electricity/base/ElectricalNetwork.java
@@ -28,26 +28,29 @@ public class ElectricalNetwork {
 
     //adds a new block to the network if it's not in it already
     public void add(IElectric be) {
-        List<Long> posList = new ArrayList<>();
-
-        members.forEach(member -> posList.add(member.getData().getId()));
-        if (posList.contains(be.getData().getId()))
-            return;
+        long id = be.getData().getId();
+		
+		for (IElectric member : this.members) {
+			if (member.getData().getId() == id) {
+				return;
+			}
+		}
         members.add(be);
     }
 
+    private float current;
     /**
      * Called when a block is removed or added to the network or when loading chunks the network is in
      * Sets up blocks in the network
      */
     public void updateNetwork() {
-
+        boolean currentValid = false;
         int maxVoltage = 0;
         float resistance = 0;
         int powerGeneration = 0;
 
 
-        /**
+        /*
          *  Phase I:
          *  1) gives each blocks the networks id
          *  2) finds the highest voltage generated
@@ -56,16 +59,12 @@ public class ElectricalNetwork {
          */
         for (IElectric member : members) {
             member.getData().notEnoughPower = false;
-
             member.getData().highestCurrent = 0;
-
             //    member.getLevelAccessor().setBlock(member.getBlockPos().above(2), Blocks.GOLD_BLOCK.defaultBlockState(),2);
-
             maxVoltage = Math.max(member.voltageGeneration(), maxVoltage);
-
             powerGeneration = (int) (powerGeneration + member.powerGeneration());
         }
-        /**
+        /*
          *  Phase II:
          * 1) informs blocks about voltage and power change
          * 2) sets network's resistance
@@ -75,31 +74,30 @@ public class ElectricalNetwork {
         if (!members.isEmpty()) {
 
             for (IElectric member : list) {
-
                 int oldVoltage = member.getData().getVoltage();
                 float oldPower = member.getPowerUsage();
                 member.getData().voltageSupply = maxVoltage;
                 member.setVoltage(maxVoltage);
                 member.getData().setVoltageNextTick = true;
-
                 member.getData().networkPowerGeneration = powerGeneration;
-
                 member.onNetworkChanged(oldVoltage, oldPower);
                 //if (member.resistance() != 0)
                 //    resistance += 1f / member.resistance();
-
             }
         }
-        /**
+
+        /*
          * Phase III:
          * 1) sets the current of wires
          * 2) informs subnetworks
          */
         for (IElectric member : members) {
-
             if (member.resistance() == 0) {
-
-                member.getData().highestCurrent = getCableCurrent(member);
+                if (!currentValid) {
+                    this.current = ElectricalNetwork.getCableCurrent(member);
+                    currentValid = true;
+                }
+                member.getData().highestCurrent = this.current;
             }
             if (member instanceof VoltageAlteringBlockEntity be) {
                 be.updateInFront();
@@ -107,16 +105,14 @@ public class ElectricalNetwork {
            // if (resistance != 0) {
            //     member.setNetworkResistance(1f / resistance);
            // } else member.setNetworkResistance(0);
-
         }
-        /**
+        /*
          * Phase IV:
          * 1) stops the network from functioning if it consumes more power than it creates
          */
         if (!members.isEmpty())
             members.getFirst().doActionNextTick(i-> members.getFirst().recalculateNetworkResistance());
         handleInsufficientPower();
-
     }